small changes
For now student-nav is useless -> commented out. tutor-appointments had some residual CSS linked that we aren't using, that reference is gone now. The tutor-register page had the logged in tutor-nav showing and we don't want that. Clicking the logo on the student page now brings you to the splash page instead of refreshing, because of the lack of the student-nav with the button that used to do that. Now returning to the splash page from non-tutor pages has a consistent method. Adjusted the hover color for the student portal button to match the darkness of the tutor button better.